chore: rename remaining plugin folders to hyphen
